Refactor luigi script into multiple parts
